Uses a pure rust implementation of SHA256 instead of the rust/c/assembler mix from the `ring` crate. This improves compatibility with e.g. webassembly and shows better runtime performance.

On my machine, this slightly reduces build time. According to the package benchmarks, the new implementation is also a bit faster (might be worth double-checking):
- content-id-text: 4.5951 -> 4.5082 ms
- content-id-image: 19.466 -> 19.211 ms
- data-id: 1.1213 -> 1.1205 ms
- instance-id: 476.80 -> 273.74 μs